BLINK(USSD)
What's the problem?

Part of the World Health Organization's guidance on limiting further spread of COVID-19 is to practice social distancing. As a result, schools in most affected areas are taking precautionary measures by closing their facilities. With school-aged children at home for an indeterminate amount of time, keeping them engaged, entertained, and on top of their education is important.

How can technology help?

Men and women has to understand how to get rid of the futher spread of this virus no matter the kind of device is beign accessible to them. And also be able to conduct a quick check for themselves and put a call through to appropraite bodies involved in the treatment of these cases.

The idea

USSD tracker system can go along way in assiting both those in the remote environment understand the common cause and ways of spread of this virus and appropriate ways of preventing it. They could also conduct a quick check for themselves through the USSD system and put WHO recommended numbers accross for quick reponse.